Making the program a bit more modular. - This makes it so I can add more functionality. - This will also allow for a more reasonable UI implementation.
